backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
flask-cors==6.0.1
Flask-Mail==0.10.0
psycopg2==2.9.10
python-dotenv==1.1.0
Faker==37.4.0
PyPDF2==3.0.1
python-docx==1.2.0
pandas==2.3.0
pydantic==2.10.6
